requirements.txt
torch
torchvision
opencv-python
fastapi
uvicorn[standard]
streamlit
pillow
numpy
requests
